DCS-94 Splitting incident and statement service
