+++ server/routes/incidents.test.js
@@ -3,12 +3,11 @@ const { appSetup } = require('./testutils/appSetup')
 const createRouter = require('./incidents')
 const { authenticationMiddleware } = require('./testutils/mockAuthentication')
 
-const incidentService = {
+const statementService = {
   getStatementsForUser: () => [{ id: 1, booking_id: 2, created_date: '12/12/2018', user_id: 'ITAG_USER' }],
   getStatement: () => ({ id: 1, booking_id: 2, created_date: '12/12/2018', user_id: 'ITAG_USER' }),
   submitStatement: jest.fn(),
   saveStatement: jest.fn(),
-  processUserInput: () => ({}),
   validateSavedStatement: jest.fn(),
 }
 
@@ -16,12 +15,12 @@ const offenderService = {
   getOffenderNames: () => [],
   getOffenderDetails: () => ({}),
 }
-const route = createRouter({ authenticationMiddleware, incidentService, offenderService })
+const route = createRouter({ authenticationMiddleware, statementService, offenderService })
 
 let app
 
 beforeEach(() => {
-  incidentService.validateSavedStatement.mockReturnValue([])
+  statementService.validateSavedStatement.mockReturnValue([])
 
   app = appSetup(route)
 })
@@ -94,7 +93,7 @@ describe('POST /incidents/:incidentId/statement/confirm', () => {
       .expect('Location', '/incidents/-1/statement/submitted'))
 
   it('confirmed submit redirects due to form not being complete', () => {
-    incidentService.validateSavedStatement.mockReturnValue([{ href: '#field', text: 'An error' }])
+    statementService.validateSavedStatement.mockReturnValue([{ href: '#field', text: 'An error' }])
     return request(app)
       .post('/incidents/-1/statement/confirm')
       .expect(302)
