+++ server/services/incidentService.js
@@ -1,8 +1,5 @@
 const logger = require('../../log.js')
 const { isNilOrEmpty } = require('../utils/utils')
-const { StatementStatus } = require('../config/types')
-const statementConfig = require('../config/statement')
-const { validate } = require('../utils/fieldValidation')
 
 module.exports = function createIncidentService({ incidentClient, elite2ClientBuilder }) {
   function getCurrentDraftIncident(userId, bookingId) {
@@ -68,48 +65,14 @@ module.exports = function createIncidentService({ incidentClient, elite2ClientBu
     }
   }
 
-  const getStatementsForUser = async (userId, status) => {
-    const data = await incidentClient.getStatementsForUser(userId, status)
-    return data.rows
-  }
-
-  const getStatement = async (userId, incidentId, status) => {
-    const statement = await incidentClient.getStatement(userId, incidentId, status)
-    if (!statement) {
-      throw new Error(`Incident: '${incidentId}' does not exist`)
-    }
-    return statement
-  }
-
-  const validateSavedStatement = async (username, incidentId) => {
-    const statement = await getStatement(username, incidentId, StatementStatus.PENDING)
-    const errors = statementConfig.validate ? validate(statementConfig.fields, statement, true) : []
-    return errors
-  }
-
   const getInvolvedStaff = incidentId => {
     return incidentClient.getInvolvedStaff(incidentId)
-  }
-
-  const saveStatement = (userId, incidentId, statement) => {
-    logger.info(`Saving statement for user: ${userId} and incident: ${incidentId}`)
-    return incidentClient.saveStatement(userId, incidentId, statement)
-  }
-
-  const submitStatement = (userId, incidentId) => {
-    logger.info(`Submitting statement for user: ${userId} and incident: ${incidentId}`)
-    return incidentClient.submitStatement(userId, incidentId)
   }
 
   return {
     getCurrentDraftIncident,
     update,
     submitForm,
-    getStatement,
-    getStatementsForUser,
     getInvolvedStaff,
-    saveStatement,
-    submitStatement,
-    validateSavedStatement,
   }
 }
